+++ src/sys/mem/borrow/mod.rs
@@ -4,6 +4,7 @@
 //
 
 mod maybe; // MaybeOwned
+mod ownership; // Ownership, BackingChoice
 mod reexports;
 
 crate::items! { // structural access: _mods, _all
@@ -11,9 +12,9 @@ crate::items! { // structural access: _mods, _all
     pub use _mods::*;
 
     mod _mods { #![allow(unused)]
-        pub use super::{maybe::*, reexports::*};
+        pub use super::{maybe::*, ownership::*, reexports::*};
     }
-    pub(super) mod _all {
+    pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]
         pub use super::_mods::*;
     }